requirements.txt
amqp==2.5.2
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.2
click==7.1.2
Django==3.0.6
djangorestframework==3.11.0
flake8==3.8.1
kombu==4.6.8
mccabe==0.6.1
pathspec==0.8.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.5.14
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
vine==1.3.0
